backend/requirements.txt
fastapi
uvicorn
sqlalchemy
psycopg2-binary
pydantic
pydantic-settings
python-jose[cryptography]
passlib[bcrypt]
python-multipart
requests
beautifulsoup4
scikit-learn
joblib
numpy
pandas
websockets
python-dotenv
email-validator
PyPDF2
